Fix RuntimeWarning when running python -m envdir
