+++ .gitignore
@@ -7,3 +7,4 @@ htmlcov
 .coverage
 docs/_build
 dist/
+**/*.pyc
